edited formatting and added gradle
