+++ build.gradle
@@ -28,11 +28,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("Doopies")
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "doopies"
     archiveClassifier = null
 }
