+++ src/main/java/Doopies/Exception/EmptyDescriptionException.java
@@ -1,4 +1,4 @@
-package Doopies.Exception;
+package doopies.exception;
 
 public class EmptyDescriptionException extends Exception {
     public EmptyDescriptionException(String message) {
