--- src/main/java/Doopies/Exception/IncorrectArgumentsException.java
@@ -1,7 +1,0 @@
-package Doopies.Exception;
-
-public class IncorrectArgumentsException extends Exception {
-    public IncorrectArgumentsException(String message) {
-        super(message);
-    }
-}
