+++ src/main/java/Doopies/Exception/InvalidTaskTypeException.java
@@ -1,4 +1,4 @@
-package Doopies.Exception;
+package doopies.exception;
 
 public class InvalidTaskTypeException extends Exception {
     public InvalidTaskTypeException(String message) {
