+++ src/main/java/Doopies/command/UnknownCommand.java
@@ -1,8 +1,8 @@
-package Doopies.command;
+package doopies.command;
 
-import Doopies.Exception.UnknownCommandException;
-import Doopies.notebook.Notebook;
-import Doopies.userInterface.Ui;
+import doopies.exception.UnknownCommandException;
+import doopies.notebook.Notebook;
+import doopies.userinterface.Ui;
 
 public class UnknownCommand extends Command {
     public UnknownCommand() {
