--- src/main/java/Doopies/util/Triple.java
@@ -1,30 +1,0 @@
-package Doopies.util;
-
-public class Triple<A, B, C> {
-    private final A a;
-    private final B b;
-    private final C c;
-
-    public Triple(A a, B b, C c) {
-        this.a = a;
-        this.b = b;
-        this.c = c;
-    }
-
-    public A getFirst() {
-        return this.a;
-    }
-
-    public B getSecond() {
-        return this.b;
-    }
-
-    public C getThird() {
-        return this.c;
-    }
-
-    @Override
-    public String toString() {
-        return String.format("(%s, %s, %s)", this.a, this.b, this.c);
-    }
-}
