+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm Doopies
+Hello! I'm doopies
 What can I do for you?
 ____________________________________________________________
